# Tritonv2: faster `triton` dequant kernel and refactored `QuantLinear`