--- a/auto_gptq/modeling/auto.py
+++ b/auto_gptq/modeling/auto.py
@@ -1,29 +1,29 @@
 from inspect import signature
 from typing import Dict, Optional, Union
 
-from ._base import BaseQuantizeConfig, BaseGPTQForCausalLM
+from ._base import BaseGPTQForCausalLM, BaseQuantizeConfig
 from ._utils import check_and_get_model_type
+from .baichuan import BaiChuanGPTQForCausalLM
 from .bloom import BloomGPTQForCausalLM
 from .codegen import CodeGenGPTQForCausalLM
+from .decilm import DeciLMGPTQForCausalLM
+from .gpt2 import GPT2GPTQForCausalLM
+from .gpt_bigcode import GPTBigCodeGPTQForCausalLM
 from .gpt_neox import GPTNeoXGPTQForCausalLM
 from .gptj import GPTJGPTQForCausalLM
-from .gpt2 import GPT2GPTQForCausalLM
+from .internlm import InternLMGPTQForCausalLM
 from .llama import LlamaGPTQForCausalLM
+from .longllama import LongLlamaGPTQForCausalLM
+from .mistral import MistralGPTQForCausalLM
+from .mixtral import MixtralGPTQForCausalLM
 from .moss import MOSSGPTQForCausalLM
 from .opt import OPTGPTQForCausalLM
-from .rw import RWGPTQForCausalLM
-from .gpt_bigcode import GPTBigCodeGPTQForCausalLM
-from .baichuan import BaiChuanGPTQForCausalLM
-from .internlm import InternLMGPTQForCausalLM
 from .qwen import QwenGPTQForCausalLM
-from .mistral import MistralGPTQForCausalLM
-from .yi import YiGPTQForCausalLM
-from .xverse import XverseGPTQForCausalLM
-from .decilm import DeciLMGPTQForCausalLM
-from .stablelmepoch import StableLMEpochGPTQForCausalLM
-from .mixtral import MixtralGPTQForCausalLM
 from .qwen2 import Qwen2GPTQForCausalLM
-from .longllama import LongLlamaGPTQForCausalLM
+from .rw import RWGPTQForCausalLM
+from .stablelmepoch import StableLMEpochGPTQForCausalLM
+from .xverse import XverseGPTQForCausalLM
+from .yi import YiGPTQForCausalLM
 
 GPTQ_CAUSAL_LM_MODEL_MAP = {
     "bloom": BloomGPTQForCausalLM,
@@ -100,6 +100,7 @@ class AutoGPTQForCausalLM:
         trainable: bool = False,
         disable_exllama: Optional[bool] = None,
         disable_exllamav2: bool = False,
+        use_tritonv2: bool = False,
         **kwargs
     ) -> BaseGPTQForCausalLM:
         # If disable_exllamav2 is True, we want to fall back on the exllama kernel and not the cuda/cuda_old ones.
@@ -108,7 +109,7 @@ class AutoGPTQForCausalLM:
                 disable_exllama = False
             else:
                 disable_exllama = True
-        
+
         model_type = check_and_get_model_type(model_name_or_path, trust_remote_code)
         quant_func = GPTQ_CAUSAL_LM_MODEL_MAP[model_type].from_quantized
         # A static list of kwargs needed for huggingface_hub
@@ -122,12 +123,13 @@ class AutoGPTQForCausalLM:
             "revision",
             "subfolder",
             "_raise_exceptions_for_missing_entries",
-            "_commit_hash"
+            "_commit_hash",
         ]
         # TODO: do we need this filtering of kwargs? @PanQiWei is there a reason we can't just pass all kwargs?
         keywords = {
             key: kwargs[key]
-            for key in list(signature(quant_func).parameters.keys()) + huggingface_kwargs
+            for key in list(signature(quant_func).parameters.keys())
+            + huggingface_kwargs
             if key in kwargs
         }
         return quant_func(
@@ -148,6 +150,7 @@ class AutoGPTQForCausalLM:
             trainable=trainable,
             disable_exllama=disable_exllama,
             disable_exllamav2=disable_exllamav2,
+            use_tritonv2=use_tritonv2,
             **keywords
         )