## Faster Triton Dequant Kernel and QuantLinear

Add a `triton` `Quantlinear` layer (`qlinear_tritonv2`) that uses a refactored, faster dequantization kernel.

The new [kernel](https://github.com/jeromeku/AutoGPTQ/blob/e218ab5dda4abff14f3fdd40533bb4ff69b18986/auto_gptq/nn_modules/triton_utils/dequant.py#L21) separates dequantization from matmul. When incorporated into a `QuantLinear` [layer](https://github.com/jeromeku/AutoGPTQ/blob/tritonv2/auto_gptq/nn_modules/qlinear/qlinear_tritonv2.py), performance is on par with `exllama` and `exllamav2` kernels in both layer-level and model-level benchmarks.

This kernel should enable faster `peft`-style fine-tuning of pre-quantized `GPTQ` models (e.g., see this [PR](https://github.com/unslothai/unsloth/pull/141)).

### Correctness

Comparing the outputs of `qlinear_tritonv2.QuantLinear` with `qlinear_cuda_old.QuantLinear` for the `self_attn` and `mlp` layers of pre-quantized `TheBloke/Llama-2-7B-GPTQ` gives errors similar to those of the original `qlinear_triton.QuantLinear` layer.

- `tritonv2` against `cuda_old`

```
    Layer 0 self_attn q_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.00266265869140625
    Layer 0 self_attn k_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.0026721954345703125
    Layer 0 self_attn v_proj: diff=Maxdiff: 0.001953125, Mean relative diff: 0.00232696533203125
    Layer 0 self_attn o_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.0027675628662109375

    Layer 0 mlp gate_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.0025005340576171875
    Layer 0 mlp up_proj: diff=Maxdiff: 0.00390625, Mean relative diff: 0.0029392242431640625
    Layer 0 mlp down_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.0020275115966796875
```

- `triton` vs `cuda_old`

```
    Layer 0 self_attn q_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.0026645660400390625
    Layer 0 self_attn k_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.0026702880859375
    Layer 0 self_attn v_proj: diff=Maxdiff: 0.001953125, Mean relative diff: 0.00232696533203125
    Layer 0 self_attn o_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.0027637481689453125

    Layer 0 mlp gate_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.0025005340576171875
    Layer 0 mlp up_proj: diff=Maxdiff: 0.00390625, Mean relative diff: 0.0029392242431640625
    Layer 0 mlp down_proj: diff=Maxdiff: 0.0078125, Mean relative diff: 0.002025604248046875
```

**Repro**

```python
python benchmark_tritonv2_integration.py --test --test_kernel=tritonv2
python benchmark_tritonv2_integration.py --test --test_kernel=triton
```

### Qlinear Benchmarks

Benchmarking qlinear_tritonv2.QuantLinear against existing kernels using quantized state of `q_proj` of `layer 0` of `TheBloke/Llama-2-7b-GPTQ`:

_Median times (ms)_

```
   seqlen  qlinear_cuda_old  qlinear_cuda  qlinear_triton  qlinear_tritonv2  qlinear_exllama  qlinear_exllamav2
0   128.0          0.564224      0.771072        0.171008          0.124928         0.158720           0.161792
1   256.0          0.590848      0.806912        0.330752          0.160768         0.141312           0.144384
2   512.0          0.646144      0.861184        0.547840          0.210992         0.226304           0.228352
3  1024.0          0.800480      1.014784        0.952320          0.366480         0.347136           0.349184
4  2048.0          1.105408      1.324032        1.761856          0.684576         0.642048           0.637952
```

**Repro**

```python
python benchmark_qlinear.py
```

### Integration Benchmarks

We also see speedups running the model end-to-end with `qlinear_tritonv2.QuantLinear` using `torch.utils.benchmark.Timer.adaptive_autorange` to benchmark the forward pass of a model (results for `TheBloke/Llama-2-7b-GPTQ` shown below).

- `qlinear_cuda`

```
 Median: 816.35 ms
 IQR:    0.65 ms (815.93 to 816.58)
 4 measurements, 1 runs per measurement, 8 threads
```

- `qlinear_triton`

```
 Median: 1.46 s
  IQR:    0.00 s (1.46 to 1.47)
  4 measurements, 1 runs per measurement, 8 threads
```

- `qlinear_tritonv2`

```
  Median: 657.80 ms
  IQR:    1.53 ms (657.15 to 658.68)
  4 measurements, 1 runs per measurement, 8 threads
```

- `qlinear_exllama`

```
  Median: 633.93 ms
  IQR:    1.97 ms (633.19 to 635.16)
  4 measurements, 1 runs per measurement, 8 threads
```

- `qlinear_exllamav2`

```
  Median: 634.38 ms
  IQR:    1.64 ms (633.66 to 635.30)
  4 measurements, 1 runs per measurement, 8 threads

```

**Repro**

```python
python benchmark_tritonv2_integration.py --benchmark_kernel=tritonv2
```

Replace `benchmark_kernel` with desired kernel: `default`, `triton`, `tritonv2`, `exllama`, `exllamav2`.

Alternatively, run the following to benchmark all kernels and save results to `{kernel}_bench.txt`

```
chmod +x run_triton_benchmark.sh
./run_triton_benchmark.sh
```
@fxmarty