@fxmarty 2024-02-12
Thank you @jeromeku, will have a look shortly. Did you write this one yourself?