@fxmarty 2024-02-13
by the way https://pytorch.org/blog/accelerating-triton/ may be interesting to you.